pyproject.toml
[tool.poetry]
name = "foodcourt"
version = "0.1.0"
description = "Social network for food lovers"
authors = ["Damn_Imposter <33062172+Impo5ter@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
Pillow = "^8.2.0"
psycopg2 = "^2.8.6"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.6.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
